feat: add convertedamount filter

+++ src/internal/common/fetcherresponse/fetcherresponse.go
@@ -3,5 +3,5 @@ package fetcherresponse
 type FetcherResponse struct {
 	Date string  `json:"date"`
 	Base string  `json:"base"`
-	Rate float64 `json:"rate"`
+	Rate float32 `json:"rate"`
 }
